update to 1.2.7 interface changes
- Add key --height Set height of terminal window. Usage: --height=xx
(where xx >= 30)
- Height of top section of split screen dependence from count of cuda
device or choosed by -d key.
- Some fixes for output info string

+++ hw_nvidia.h
@@ -9,17 +9,17 @@
 
 // Retrieve the temperature (in degrees) of a Nvidia GPU. -1 if not available
 int nw_nvidia_init();
-DWORD hw_nvidia_gettemperature( DWORD dwGPUIndex );
-DWORD hw_nvidia_DynamicPstateInfoEx( DWORD dwGPUIndex );
-DWORD hw_nvidia_Pstate20 (DWORD dwGPUIndex);
-DWORD hw_nvidia_memory (DWORD dwGPUIndex);
-DWORD hw_nvidia_memory_prc (DWORD dwGPUIndex);
-DWORD hw_nvidia_memory_util_prc( DWORD dwGPUIndex );
-DWORD hw_nvidia_clock (DWORD dwGPUIndex);
-DWORD hw_nvidia_clockMemory (DWORD dwGPUIndex);
-DWORD hw_nvidia_voltage (DWORD dwGPUIndex);
-DWORD hw_nvidia_cooler (DWORD dwGPUIndex);
-DWORD hw_nvidia_fan (DWORD dwGPUIndex);
-DWORD get_bus_id(DWORD dwGPUindex);
-DWORD hw_nvidia_version();
-DWORD get_bus_ids();
+INT32 hw_nvidia_gettemperature( DWORD dwGPUIndex );
+ULONG hw_nvidia_DynamicPstateInfoEx( DWORD dwGPUIndex );
+ULONG hw_nvidia_Pstate20 (DWORD dwGPUIndex);
+ULONG hw_nvidia_memory (DWORD dwGPUIndex);
+ULONG hw_nvidia_memory_prc (DWORD dwGPUIndex);
+ULONG hw_nvidia_memory_util_prc( DWORD dwGPUIndex );
+ULONG hw_nvidia_clock (DWORD dwGPUIndex);
+ULONG hw_nvidia_clockMemory (DWORD dwGPUIndex);
+ULONG hw_nvidia_voltage (DWORD dwGPUIndex);
+ULONG hw_nvidia_cooler (DWORD dwGPUIndex);
+ULONG hw_nvidia_fan (DWORD dwGPUIndex);
+ULONG get_bus_id(DWORD dwGPUindex);
+ULONG hw_nvidia_version();
+ULONG get_bus_ids();
